Use workaround for `std::process::Command` segfaults and clang errors
Before this commit, the program would sometimes segfault in
`std::process::Command::arg` and sometimes report a `clang: error: no
input files` error. It feels like these could be alternate forms of
the same issue.

On a whim, I tried invoking shell commands like this. It succeeds,
seemingly reliably. I'm not sure what to make of that, and it seems
likely that this avoids some sort of race condition by virtue of
invoking a child process from the child process.

A good start for debugging the underlying cause would be to revert
this fix and watch the syscall made when clang errors.