Fix interpreter functions-calling-functions
We allow for function names to be reused. Functions calling functions
should call functions that are defined when they are defined. Later
redefinitions of that function should be ignored. The interpreter was
not handling that logic because it did not track which were the
correct functions.

```
f(a, b) = a + b;
g(x) = f(x, x + 1);
f(a) = a;
n = g(1);
```

Until now, `n` would invoke `g` and `g` would attempt (and fail) to
use the second definition of `f`. With this patch, it should correctly
use the first definition of `f`.