Replace single-use functions with blocks on `main`
For simple programs, this results in much more understandable LLVM IR.
For more complex ones, for instance with lots of `match` expressions,
it's probably less good. The ideal might be to have `run` remain a
separate function, but I think until this code is more cleaned up it's
best to tightly integrate everything.