Support for evaluating recursive functions
While starting to write a README, I discovered you couldn't write
recursive functions. With the lack of iteration, that drastically
limits what the language is capable of. Thus the need to implement
recursion.

The arbitrary program generation does not yet generate programs with
recursion. For now it is too hard to avoid run-time stack overflows.